Issue 10 I forgot to put the main and started working on the issue 11 - This one has main and works properly

+++ raster.h
@@ -3,7 +3,7 @@
 
 typedef unsigned char UINT8;
 typedef unsigned short UINT16;
-typedef unsigned int UINT32;
+typedef unsigned long UINT32;
 
 /*----- Function: clear_screen -----
 v
